+++ common/include/math_utils.hpp
@@ -3,6 +3,8 @@
 #include <algorithm>
 #include <numeric>
 #include <iterator>
+#include <vector>
+#include <cmath>
 
 namespace math
 {
@@ -12,8 +14,8 @@ auto average(
     const NumberIter end,
     Pred pred)
 {
-    const double sum = std::accumulate(beg, end, 0.0, pred);
-    const double size = std::distance(beg, end);
+    const double sum = static_cast<double>(std::accumulate(beg, end, 0.0, pred));
+    const size_t size = std::distance(beg, end);
     if (!size)
     {
         return 0.0;
@@ -27,7 +29,7 @@ auto average(
     const NumberIter beg,
     const NumberIter end)
 {
-    return average(beg, end, [](float sum, auto& item) { return sum + item; });
+    return average(beg, end, [](double sum, auto& item) { return sum + item; });
 }
 
 template <typename NumberIter, typename BinSize>
@@ -46,13 +48,13 @@ auto bin(
 
     const auto min_max = std::minmax_element(beg, end);
 
-    const size_t num_of_bins = (*min_max.second - *min_max.first + bin_size)/bin_size;
+    const size_t num_of_bins = static_cast<size_t>((*min_max.second - *min_max.first + bin_size)/bin_size);
 
     bins.resize(num_of_bins);
 
     for (auto it = beg; it != end; ++it)
     {
-        const size_t dest_bin = (*it - *min_max.first)/bin_size;
+        const size_t dest_bin = static_cast<size_t>((*it - *min_max.first)/bin_size);
         bins[dest_bin].push_back(*it);
     }
